% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/China_covid19.R
\docType{data}
\name{China_covid19}
\alias{China_covid19}
\title{The situation of Coronavirus Cases in China}
\format{
the data set contains 8349 observations and 8 variables.
\itemize{
\item \strong{Province/Stage}: The province of state in China
\item \strong{Lat}: The latitude point of the province or state
\item \strong{Long}: The longitude point of the province or state
\item \strong{Date}: The date in formate of y/m/d
\item \strong{Confirmed}: Daily confirmed cases
\item \strong{Deaths}: Daily deaths cases
\item \strong{Recovered}: Daily recovered cases
\item \strong{Month}: The month of the day
}
}
\source{
\href{https://data.humdata.org/dataset/novel-coronavirus-2019-ncov-cases}{HDX}
}
\usage{
China_covid19
}
\description{
This dataset contains the daily cases of covid-19 with all provinces in China from 1/22/2020 to 9/30/2020,
including Confirmed, Deaths and Recovered cases.
}
\keyword{datasets}
