% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cases_province.R
\name{plot_cases_province}
\alias{plot_cases_province}
\title{Plot the bar chart of all the provinces with daily cases in a dateset}
\usage{
plot_cases_province(data)
}
\arguments{
\item{data}{A dataset containing \code{Province/State} and \code{Type} of \code{Cases} in columns.}
}
\value{
A ggplot bar chart of COVID-19 with \verb{type of cases} in three colors among all \code{provinces} on a specific day.
}
\description{
Creates a bar chart to present all types of daily cases in all provinces.
}
\examples{
plot_cases_province(China_type)

}
