% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_provinceInput.R
\name{select_provinceInput}
\alias{select_provinceInput}
\title{Selection of provinces for users in app}
\usage{
select_provinceInput(data)
}
\arguments{
\item{data}{A dataset containing \code{Province/State} in columns.}
}
\value{
A shiny drop down menu with all provinces in data set.
}
\description{
Creates selection input in shiny package for users to choose which province they want to check with.
}
